Snapcast u pn p (#1)
* bug fixess, getops and module additions

- each function is now callable on an individual level with command line flags
- if statement for debian version checks.
- inclusion of snapcast installation
- inclusion of UPnP client installation
- README updated to include new modules.

new version of shairport and settings defaults changed to be a bit more useful and a little less deafening.

* Update install.sh

fix getops arguments and include fix lost with merge

* Update install.sh

updated to have clearer communication of progress (banners and colouring)

Added a few 'basic os' setup commands. Specifically netselect-apt

the function update_latest is not set to execute. I'm not sure it's a good idea to force it on people just yet.

* Update install.sh

- heading text returns to normal colour after writing a message.
- Fix mistakes with function and variable calls.
- enabling bt-agent '--now' flag removed due to crashes
- updated Get Opts help message.
- moved OS verification and netselect to the standard installation path. With yes no flag.

+++ install.sh
@@ -3,7 +3,7 @@
 set -e
 
 NQPTP_VERSION="1.2.4"
-SHAIRPORT_SYNC_VERSION="4.3.2"
+SHAIRPORT_SYNC_VERSION="4.3.6"
 TMP_DIR=""
 
 cleanup() {
@@ -12,42 +12,135 @@ cleanup() {
     fi
 }
 
+log_green() {
+  local text="$1"
+  GREEN="\033[0;32m"
+  NORMAL=$(tput sgr0)
+  printf "${GREEN} ${text}${NORMAL}\r\n"
+}
+
+log_red() {
+  local text="$1"
+  RED="\033[0;31m"
+  NORMAL=$(tput sgr0)
+  echo $text
+  printf "${RED} ${text}${NORMAL}\r\n"
+}
+
+banner(){
+  # Get the terminal width
+  width=$(tput cols)
+  # Print a line of '=' characters
+  printf '=%.0s' $(seq 1 $width)
+}
+
+apt_update_netselect(){
+    banner
+    read -p "Do you want to change apt (netselect-apt)? [y/N] " REPLY
+    if [[ ! "$REPLY" =~ ^(yes|y|Y)$ ]]; then return; fi
+    sudo apt-get update
+    log_green "installing netselect-apt and executing"
+    sudo apt-get install netselect-apt -y
+    sudo netselect-apt
+}
+
+update_latest(){
+    sudo apt-get update 
+    sudo apt-get upgrade -y
+}
+
 verify_os() {
     MSG="Unsupported OS: Raspberry Pi OS 12 (bookworm) is required."
 
     if [ ! -f /etc/os-release ]; then
-        echo $MSG
+        log_red $MSG
+        banner
         exit 1
     fi
 
     . /etc/os-release
 
     if [ "$ID" != "debian" ] && [ "$ID" != "raspbian" ] || [ "$VERSION_ID" != "12" ]; then
-        echo $MSG
+        log_red $MSG
+        banner
         exit 1
     fi
 }
 
 set_hostname() {
-    CURRENT_PRETTY_HOSTNAME=$(hostnamectl status --pretty)
+    if [[ -z $changeHostname ]]; then 
+      if ! $changeHostname ; then return; fi
 
-    read -p "Hostname [$(hostname)]: " HOSTNAME
-    sudo raspi-config nonint do_hostname ${HOSTNAME:-$(hostname)}
+      log_green "Device Name Settings"
+      banner
 
-    read -p "Pretty hostname [${CURRENT_PRETTY_HOSTNAME:-Raspberry Pi}]: " PRETTY_HOSTNAME
-    PRETTY_HOSTNAME="${PRETTY_HOSTNAME:-${CURRENT_PRETTY_HOSTNAME:-Raspberry Pi}}"
-    sudo hostnamectl set-hostname --pretty "$PRETTY_HOSTNAME"
+      CURRENT_PRETTY_HOSTNAME=$(hostnamectl status --pretty)
+
+      read -p "Hostname [$(hostname)]: " HOSTNAME
+      sudo raspi-config nonint do_hostname ${HOSTNAME:-$(hostname)}
+
+      read -p "Pretty hostname [${CURRENT_PRETTY_HOSTNAME:-Raspberry Pi}]: " PRETTY_HOSTNAME
+      PRETTY_HOSTNAME="${PRETTY_HOSTNAME:-${CURRENT_PRETTY_HOSTNAME:-Raspberry Pi}}"
+      sudo hostnamectl set-hostname --pretty "$PRETTY_HOSTNAME"
+    fi
+}
+
+install_snapcast(){
+    if [[ -z $snapclientInstall ]]; then 
+      read -p "Do you want to install UPnP renderer? [y/N] " REPLY
+      #https://github.com/Torgee/rpi-audio-receiver/blob/master/install-snapcast.sh
+      if [[ ! "$REPLY" =~ ^(yes|y|Y)$ ]]; then return; fi
+    fi
+    if ! $snapclientInstall; then return; fi
+
+    log_green "Installing snapcast client"
+    banner
+
+    sudo apt install --no-install-recommends -y snapclient    
+}
+
+install_UPnP_renderer(){
+    if [[ -z $UPnPRendererInstall ]]; then 
+      read -p "Do you want to install UPnP renderer? [y/N] " REPLY
+      #https://github.com/Torgee/rpi-audio-receiver/blob/master/install-upnp.sh
+      if [[ ! "$REPLY" =~ ^(yes|y|Y)$ ]]; then return; fi
+    fi
+    if ! $UPnPRendererInstall ; then return; fi
+
+    log_green "Installing UPnP renderer (gmrender-resurrect)"
+    banner
+
+    sudo apt update
+    sudo apt install -y --no-install-recommends gmediarender gstreamer1.0-plugins-base gstreamer1.0-plugins-good gstreamer1.0-plugins-ugly gstreamer1.0-alsa
+
+    LIBRESPOT_NAME="${PRETTY_HOSTNAME// /-}"
+    LIBRESPOT_NAME=${LIBRESPOT_NAME:-$(hostname)}
+    sudo tee /etc/default/gmediarender >/dev/null <<EOF
+ENABLED=1
+DAEMON_USER="nobody:audio"
+UPNP_DEVICE_NAME="${LIBRESPOT_NAME}"
+INITIAL_VOLUME_DB=0.0
+ALSA_DEVICE="sysdefault"
+EOF
+
+    sudo systemctl enable --now gmediarender
 }
 
 install_bluetooth() {
-    read -p "Do you want to install Bluetooth Audio (ALSA)? [y/N] " REPLY
-    if [[ ! "$REPLY" =~ ^(yes|y|Y)$ ]]; then return; fi
+    if [[ -z $bluetoothInstall ]]; then 
+      read -p "Do you want to install Bluetooth Audio (ALSA)? [y/N] " REPLY
+      if [[ ! "$REPLY" =~ ^(yes|y|Y)$ ]]; then return; fi
+    fi
+    if ! $bluetoothInstall ; then return; fi
 
-    # Bluetooth Audio ALSA Backend (bluez-alsa-utils)
+    log_green "Bluetooth installation"
+    banner
+
+    log_green "Bluetooth: Setting Audio ALSA Backend (bluez-alsa-utils)"
     sudo apt update
     sudo apt install -y --no-install-recommends bluez-tools bluez-alsa-utils
 
-    # Bluetooth settings
+    log_green "Bluetooth: creating basic settings"
     sudo tee /etc/bluetooth/main.conf >/dev/null <<'EOF'
 [General]
 Class = 0x200414
@@ -57,7 +150,7 @@ DiscoverableTimeout = 0
 AutoEnable=true
 EOF
 
-    # Bluetooth Agent
+    log_green "Bluetooth: configuring Agent"
     sudo tee /etc/systemd/system/bt-agent@.service >/dev/null <<'EOF'
 [Unit]
 Description=Bluetooth Agent
@@ -79,7 +172,7 @@ EOF
     sudo systemctl daemon-reload
     sudo systemctl enable bt-agent@hci0.service
 
-    # Bluetooth udev script
+    log_green "Bluetooth: installing udev script"
     sudo tee /usr/local/bin/bluetooth-udev >/dev/null <<'EOF'
 #!/bin/bash
 if [[ ! $NAME =~ ^\"([0-9A-F]{2}[:-]){5}([0-9A-F]{2})\"$ ]]; then exit 0; fi
@@ -104,11 +197,29 @@ EOF
 SUBSYSTEM=="input", GROUP="input", MODE="0660"
 KERNEL=="input[0-9]*", RUN+="/usr/local/bin/bluetooth-udev"
 EOF
+
+read -p "Do you want to configure Bluetooth A2DP volume? [y/N] " REPLY
+if [[ ! "$REPLY" =~ ^(yes|y|Y)$ ]]; then return; fi
+# Enable A2DP volume control
+    sudo mkdir -p /etc/systemd/system/bluetooth.service.d
+    sudo tee /etc/systemd/system/bluetooth.service.d/override.conf >/dev/null  <<'EOF'
+[Service]
+ExecStart=
+ExecStart=/usr/libexec/bluetooth/bluetoothd --plugin=a2dp
+EOF
+
+
 }
 
 install_shairport() {
-    read -p "Do you want to install Shairport Sync (AirPlay 2 audio player)? [y/N] " REPLY
-    if [[ ! "$REPLY" =~ ^(yes|y|Y)$ ]]; then return; fi
+    if [[ -z $shairportInstall ]]; then 
+      read -p "Do you want to install Shairport Sync (AirPlay 2 audio player)? [y/N] " REPLY
+      if [[ ! "$REPLY" =~ ^(yes|y|Y)$ ]]; then return; fi
+    fi
+    if ! $shairportInstall; then return; fi
+    
+    log_green "Installing Shairport Sync"
+    banner
 
     sudo apt update
     sudo apt install -y --no-install-recommends wget unzip autoconf automake build-essential libtool git autoconf automake libpopt-dev libconfig-dev libasound2-dev avahi-daemon libavahi-client-dev libssl-dev libsoxr-dev libplist-dev libsodium-dev libavutil-dev libavcodec-dev libavformat-dev uuid-dev libgcrypt20-dev xxd
@@ -119,7 +230,7 @@ install_shairport() {
 
     cd $TMP_DIR
 
-    # Install ALAC
+    log_green "Shairport: Install ALAC"
     wget -O alac-master.zip https://github.com/mikebrady/alac/archive/refs/heads/master.zip
     unzip alac-master.zip
     cd alac-master
@@ -131,7 +242,7 @@ install_shairport() {
     cd ..
     rm -rf alac-master
 
-    # Install NQPTP
+    log_green "Shairport: Install NQPTP"
     wget -O nqptp-${NQPTP_VERSION}.zip https://github.com/mikebrady/nqptp/archive/refs/tags/${NQPTP_VERSION}.zip
     unzip nqptp-${NQPTP_VERSION}.zip
     cd nqptp-${NQPTP_VERSION}
@@ -142,7 +253,7 @@ install_shairport() {
     cd ..
     rm -rf nqptp-${NQPTP_VERSION}
 
-    # Install Shairport Sync
+    log_green "Shairport: Install Shairport Sync"
     wget -O shairport-sync-${SHAIRPORT_SYNC_VERSION}.zip https://github.com/mikebrady/shairport-sync/archive/refs/tags/${SHAIRPORT_SYNC_VERSION}.zip
     unzip shairport-sync-${SHAIRPORT_SYNC_VERSION}.zip
     cd shairport-sync-${SHAIRPORT_SYNC_VERSION}
@@ -153,7 +264,7 @@ install_shairport() {
     cd ..
     rm -rf shairport-sync-${SHAIRPORT_SYNC_VERSION}
 
-    # Configure Shairport Sync
+    log_green "Shairport: Configure Shairport Sync"
     sudo tee /etc/shairport-sync.conf >/dev/null <<EOF
 general = {
   name = "${PRETTY_HOSTNAME:-$(hostname)}";
@@ -171,38 +282,86 @@ EOF
 }
 
 install_raspotify() {
-    read -p "Do you want to install Raspotify (Spotify Connect)? [y/N] " REPLY
-    if [[ ! "$REPLY" =~ ^(yes|y|Y)$ ]]; then return; fi
+    if [[ -z $raspotifyInstall ]]; then 
+      read -p "Do you want to install Raspotify (Spotify Connect)? [y/N] " REPLY
+      if [[ ! "$REPLY" =~ ^(yes|y|Y)$ ]]; then return; fi
+    fi
+    if ! $raspotifyInstall; then return; fi
+
+    log_green "Installing Install Raspotify"
+    banner
 
     # Install Raspotify
     curl -sL https://dtcooper.github.io/raspotify/install.sh | sh
 
-    # Configure Raspotify
+    log_green "Raspotify: Configure Raspotify"
     LIBRESPOT_NAME="${PRETTY_HOSTNAME// /-}"
     LIBRESPOT_NAME=${LIBRESPOT_NAME:-$(hostname)}
 
     sudo tee /etc/raspotify/conf >/dev/null <<EOF
-LIBRESPOT_QUIET=
-LIBRESPOT_AUTOPLAY=
-LIBRESPOT_DISABLE_AUDIO_CACHE=
-LIBRESPOT_DISABLE_CREDENTIAL_CACHE=
-LIBRESPOT_ENABLE_VOLUME_NORMALISATION=
+LIBRESPOT_QUIET=on
+LIBRESPOT_AUTOPLAY=off
+LIBRESPOT_DISABLE_AUDIO_CACHE=off
+LIBRESPOT_DISABLE_CREDENTIAL_CACHE=on
+LIBRESPOT_ENABLE_VOLUME_NORMALISATION=on
 LIBRESPOT_NAME="${LIBRESPOT_NAME}"
 LIBRESPOT_DEVICE_TYPE="avr"
 LIBRESPOT_BITRATE="320"
-LIBRESPOT_INITIAL_VOLUME="100"
+LIBRESPOT_INITIAL_VOLUME="50"
 EOF
 
+    log_green "Raspotify: daemon reload and enable"
     sudo systemctl daemon-reload
-    sudo systemctl enable raspotify
+    sudo systemctl enable --now raspotify
 }
 
 trap cleanup EXIT
 
-echo "Raspberry Pi Audio Receiver"
+log_green "Raspberry Pi Audio Receiver"
+banner
 
-verify_os
-set_hostname
-install_bluetooth
-install_shairport
-install_raspotify
+changeHostname=false
+bluetoothInstall=false
+shairportInstall=false
+raspotifyInstall=false
+UPnPRendererInstall=false 
+snapclientInstall=false 
+while getopts "nbsruc" opt; do
+  case "$opt" in
+    n) changeHostname=true ;;
+    b) bluetoothInstall=true;;
+    s) shairportInstall=true;;
+    r) raspotifyInstall=true;;
+    u) UPnPRendererInstall=true;;
+    c) snapclientInstall=true;;
+    ?) echo "script usage: $(basename $0) [-n][-b][-s][-r][-u][-c]
+      -n Change Host Name
+      -b Install bluetooth features
+      -s Install Shairplay (Airplay support)
+      -r Raspotify (Spotify support)
+      -u UPnP Render Install
+      -c Snapcast Client Install" 
+      exit 1
+      ;;
+  esac
+done
+
+if (( $OPTIND == 1 )); then
+  echo "Default option"
+  changeHostname=""
+  bluetoothInstall=""
+  shairportInstall=""
+  raspotifyInstall=""
+  UPnPRendererInstall=""
+  snapclientInstall=""
+  verify_os
+  apt_update_netselect
+fi
+
+set_hostname $changeHostname
+install_bluetooth $bluetoothInstall
+install_shairport $shairportInstall
+install_raspotify $raspotifyInstall
+install_snapcast $snapclientInstall
+install_UPnP_renderer $UPnPRendererInstall
+
